try fix x86 tests

+++ src/libhwloc.jl
@@ -255,7 +255,7 @@ mutable struct Object
     os_index::Int
     name::String
     attr::Attribute
-    mem::UInt
+    mem::Int64
 
     depth::Int
     logical_index::Int
@@ -308,7 +308,7 @@ function load(hobj::hwloc_obj_t)
 
     topo.attr = load_attr(obj.attr, topo.type_)
 
-    topo.mem = UInt(obj.total_memory)
+    topo.mem = Int64(obj.total_memory)
 
     topo.depth = obj.depth
